change of structure
