+++ src/main/java/IMVC/IController.java
@@ -1,0 +1,5 @@
+package IMVC;
+
+public interface IController {
+    boolean tryLogin(byte[] loginParams);
+}
