+++ src/main/java/IMVC/IModel.java
@@ -1,0 +1,7 @@
+package IMVC;
+
+import java.util.Map;
+
+public interface IModel {
+    boolean login(Map<String, String> params);
+}
