+++ src/main/java/IMVC/IView.java
@@ -1,0 +1,5 @@
+package IMVC;
+
+public interface IView {
+    void init();
+}
