--- src/main/java/exercise1/IMVC/IController.java
@@ -1,5 +1,0 @@
-package exercise1.IMVC;
-
-public interface IController {
-    boolean tryLogin(byte[] loginParams);
-}
