--- src/main/java/exercise1/IMVC/IModel.java
@@ -1,7 +1,0 @@
-package exercise1.IMVC;
-
-import java.util.Map;
-
-public interface IModel {
-    boolean login(Map<String, String> params);
-}
