--- src/main/java/exercise1/IMVC/IView.java
@@ -1,5 +1,0 @@
-package exercise1.IMVC;
-
-public interface IView {
-    void init();
-}
