--- src/main/java/exercise1/model/IActions/IMovable.java
@@ -1,5 +1,0 @@
-package exercise1.model.IActions;
-
-public interface IMovable {
-    void move (double x, double y);
-}
