--- src/main/java/exercise1/model/IActions/IRollable.java
@@ -1,5 +1,0 @@
-package exercise1.model.IActions;
-
-public interface IRollable {
-    void roll(double angle);
-}
