--- src/main/java/exercise1/model/IActions/IScalebale.java
@@ -1,7 +1,0 @@
-package exercise1.model.IActions;
-
-public interface IScalebale {
-    default void changeSize(double scaleFactor) {
-
-    }
-}
