+++ src/main/java/model/IActions/IMovable.java
@@ -1,0 +1,5 @@
+package model.IActions;
+
+public interface IMovable {
+    void move (double x, double y);
+}
