+++ src/main/java/model/IActions/IRollable.java
@@ -1,0 +1,5 @@
+package model.IActions;
+
+public interface IRollable {
+    void roll(double angle);
+}
