+++ src/main/java/model/IActions/IScalebale.java
@@ -1,0 +1,7 @@
+package model.IActions;
+
+public interface IScalebale {
+    default void changeSize(double scaleFactor) {
+
+    }
+}
