--- src/main/resources/out.txt
@@ -1,3 +1,0 @@
-CIRCLE (6.0, 6.0)
-TRIANGLE (1.0, 1.0) (2.0, 0.0) (0.0, 0.0)
-RECTANGLE (0.0, 1.0) (2.0, 1.0) (2.0, 0.0) (0.0, 0.0)
