--- src/main/resources/test.txt
@@ -1,3 +1,0 @@
-circle 3.0 3.0 3.0
-triangle 0.0 0.0 2.0 0.0 1.0 1.0
-rectangle 0.0 1.0 0.0 0.0 2.0 0.0 2.0 1.0
